+++ Engine/math/Mat4x4.cpp
@@ -11,7 +11,7 @@ Mat4x4::Mat4x4(const bool _isIdentityMatrix)
 	}
 }
 
-Mat4x4::Mat4x4(const double _m[16]) { std::memcpy(this->m, _m, 16 * sizeof(double)); }
+Mat4x4::Mat4x4(const float _m[16]) { std::memcpy(this->m, _m, 16 * sizeof(float)); }
 
 Mat4x4::Mat4x4(const Mat4x4& _m) { Mat4x4(_m.m); }
 
@@ -28,12 +28,12 @@ Mat4x4 Mat4x4::operator*(const Mat4x4& _m) const {
 	return result;
 }
 
-Mat4x4 Mat4x4::getRotationXMatrix(const double _rotX)
+Mat4x4 Mat4x4::getRotationXMatrix(const float _rotX)
 {
-	const double sinX = std::sin(_rotX);
-	const double cosX = std::cos(_rotX);
+	const float sinX = std::sin(_rotX);
+	const float cosX = std::cos(_rotX);
 
-	double m[16] = {
+	float m[16] = {
 		1, 0,     0,     0,
 		0, +cosX, +sinX, 0,
 		0, -sinX, +cosX, 0,
@@ -43,12 +43,12 @@ Mat4x4 Mat4x4::getRotationXMatrix(const double _rotX)
 	return Mat4x4(m);
 }
 
-Mat4x4 Mat4x4::getRotationYMatrix(const double _rotY)
+Mat4x4 Mat4x4::getRotationYMatrix(const float _rotY)
 {
-	const double sinY = std::sin(_rotY);
-	const double cosY = std::cos(_rotY);
+	const float sinY = std::sin(_rotY);
+	const float cosY = std::cos(_rotY);
 
-	double m[16] = {
+	float m[16] = {
 		cosY, 0, -sinY, 0,
 		0,    1, 0,     0,
 		sinY, 0, cosY,  0,
@@ -58,12 +58,12 @@ Mat4x4 Mat4x4::getRotationYMatrix(const double _rotY)
 	return Mat4x4(m);
 }
 
-Mat4x4 Mat4x4::getRotationZMatrix(const double _rotZ)
+Mat4x4 Mat4x4::getRotationZMatrix(const float _rotZ)
 {
-	const double sinZ = std::sin(_rotZ);
-	const double cosZ = std::cos(_rotZ);
+	const float sinZ = std::sin(_rotZ);
+	const float cosZ = std::cos(_rotZ);
 
-	double m[16] = {
+	float m[16] = {
 		+cosZ, +sinZ, 0, 0,
 		-sinZ, +cosZ, 0, 0,
 		0,     0,     1, 0,
@@ -73,12 +73,12 @@ Mat4x4 Mat4x4::getRotationZMatrix(const double _rotZ)
 	return Mat4x4(m);
 }
 
-Mat4x4 Mat4x4::getPerspectiveMatrix(const unsigned int _width, const unsigned int _height, const double _fov, const double _zNear, const double _zFar)
+Mat4x4 Mat4x4::getPerspectiveMatrix(const unsigned int _width, const unsigned int _height, const float _fov, const float _zNear, const float _zFar)
 {
-	double aspectRatio = _height / static_cast<double>(_width);
-	double a = -std::abs(1.f / std::tan(MATH::degToRad(_fov / 2.f)));
+	float aspectRatio = _height / static_cast<float>(_width);
+	float a = -std::abs(1.f / std::tan(MATH::degToRad(_fov / 2.f)));
 
-	double m[16] = {
+	float m[16] = {
 			aspectRatio * a, 0, 0,                                    0,
 			0,               a, 0,                                    0,
 			0,               0, _zFar / (_zFar - _zNear),             1,
