+++ Engine/math/Mat4x4.h
@@ -5,11 +5,11 @@
 #include "../other/pch.h"
 
 struct Mat4x4 {
-	double m[16] = { 0 }; // r * c
+	float m[16] = { 0 }; // r * c
 
 	Mat4x4(const bool _isIdentityMatrix = false);
 
-	Mat4x4(const double _m[16]);
+	Mat4x4(const float _m[16]);
 
 	Mat4x4(const Mat4x4& _m);
 
@@ -17,10 +17,10 @@ struct Mat4x4 {
 
 	Mat4x4 operator*(const Mat4x4& _m) const;
 
-	static Mat4x4 getRotationXMatrix(const double _rotX);
-	static Mat4x4 getRotationYMatrix(const double _rotY);
-	static Mat4x4 getRotationZMatrix(const double _rotZ);
-	static Mat4x4 getPerspectiveMatrix(const unsigned int _width, const unsigned int _height, const double _fov, const double _zNear, const double _zFar);
+	static Mat4x4 getRotationXMatrix(const float _rotX);
+	static Mat4x4 getRotationYMatrix(const float _rotY);
+	static Mat4x4 getRotationZMatrix(const float _rotZ);
+	static Mat4x4 getPerspectiveMatrix(const unsigned int _width, const unsigned int _height, const float _fov, const float _zNear, const float _zFar);
 };
 
 // Mat4x4 Functions
