+++ Engine/math/Vec3.cpp
@@ -3,90 +3,90 @@
 // Vector Object
 
 Vec3::Vec3() : x(0), y(0), z(0), w(0) {}
-Vec3::Vec3(const double _n) : x(_n), y(_n), z(_n), w(_n) {}
-Vec3::Vec3(const double _x, const double _y) : x(_x), y(_y), z(0), w(0) {}
-Vec3::Vec3(const double _x, const double _y, const double _z) : x(_x), y(_y), z(_z), w(0) {}
-Vec3::Vec3(const double _x, const double _y, const double _z, const double _w) : x(_x), y(_y), z(_z), w(_w) {}
+Vec3::Vec3(const float _n) : x(_n), y(_n), z(_n), w(_n) {}
+Vec3::Vec3(const float _x, const float _y) : x(_x), y(_y), z(0), w(0) {}
+Vec3::Vec3(const float _x, const float _y, const float _z) : x(_x), y(_y), z(_z), w(0) {}
+Vec3::Vec3(const float _x, const float _y, const float _z, const float _w) : x(_x), y(_y), z(_z), w(_w) {}
 Vec3::Vec3(const Vec3& _v) : x(_v.x), y(_v.y), z(_v.z), w(_v.w) {}
 
-void Vec3::add(const double _x, const double _y, const double _z, const double _w, const bool _useW)
+void Vec3::add(const float _x, const float _y, const float _z, const float _w, const bool _useW)
 {
 	this->x += _x; this->y += _y; this->z += _z;
 
 	if (_useW) this->w += _w;
 }
 
-void Vec3::add(const double _x, const double _y, const double _z) { this->x += _x; this->y += _y; this->z += _z; }
-void Vec3::add(const double _x, const double _y) { this->x += _x; this->y += _y; }
+void Vec3::add(const float _x, const float _y, const float _z) { this->x += _x; this->y += _y; this->z += _z; }
+void Vec3::add(const float _x, const float _y) { this->x += _x; this->y += _y; }
 void Vec3::add(const Vec3& _v, const bool _useW) { this->add(_v.x, _v.y, _v.z, _v.w, _useW); }
-void Vec3::add(const double _n, const bool _useW) { this->add(_n, _n, _n, _n, _useW); }
+void Vec3::add(const float _n, const bool _useW) { this->add(_n, _n, _n, _n, _useW); }
 void Vec3::operator+=(const Vec3& _a) { this->add(_a); }
-void Vec3::operator+=(const double _a) { this->add(_a); }
+void Vec3::operator+=(const float _a) { this->add(_a); }
 Vec3 Vec3::operator+(const Vec3& _a) const { Vec3 result = this->copy(); result.add(_a); return result; }
-Vec3 Vec3::operator+(const double _a)   const { Vec3 result = this->copy(); result.add(_a); return result; }
+Vec3 Vec3::operator+(const float _a)   const { Vec3 result = this->copy(); result.add(_a); return result; }
 
 
 
-void Vec3::sub(const double _x, const double _y, const double _z, const double _w, const bool _useW)
+void Vec3::sub(const float _x, const float _y, const float _z, const float _w, const bool _useW)
 {
 	this->x -= _x; this->y -= _y; this->z -= _z;
 
 	if (_useW) this->w -= _w;
 }
 
-void Vec3::sub(const double _x, const double _y, const double _z) { this->x -= _x; this->y -= _y; this->z -= _z; }
-void Vec3::sub(const double _x, const double _y) { this->x -= _x; this->y -= _y; }
+void Vec3::sub(const float _x, const float _y, const float _z) { this->x -= _x; this->y -= _y; this->z -= _z; }
+void Vec3::sub(const float _x, const float _y) { this->x -= _x; this->y -= _y; }
 void Vec3::sub(const Vec3& _v, const bool _useW) { this->sub(_v.x, _v.y, _v.z, _v.w, _useW); }
-void Vec3::sub(const double _n, const bool _useW) { this->sub(_n, _n, _n, _n, _useW); }
+void Vec3::sub(const float _n, const bool _useW) { this->sub(_n, _n, _n, _n, _useW); }
 void Vec3::operator-=(const Vec3& _a) { this->sub(_a); }
-void Vec3::operator-=(const double _a) { this->sub(_a); }
+void Vec3::operator-=(const float _a) { this->sub(_a); }
 Vec3 Vec3::operator-(const Vec3& _a) const { Vec3 result = this->copy(); result.sub(_a); return result; }
-Vec3 Vec3::operator-(const double _a)   const { Vec3 result = this->copy(); result.sub(_a); return result; }
+Vec3 Vec3::operator-(const float _a)   const { Vec3 result = this->copy(); result.sub(_a); return result; }
 
 
-void Vec3::mul(const double _x, const double _y, const double _z, const double _w, const bool _useW)
+void Vec3::mul(const float _x, const float _y, const float _z, const float _w, const bool _useW)
 {
 	this->x *= _x; this->y *= _y; this->z *= _z;
 
 	if (_useW) this->w *= _w;
 }
 
-void Vec3::mul(const double _x, const double _y, const double _z) { this->x *= _x; this->y *= _y; this->z *= _z; }
-void Vec3::mul(const double _x, const double _y) { this->x *= _x; this->y *= _y; }
+void Vec3::mul(const float _x, const float _y, const float _z) { this->x *= _x; this->y *= _y; this->z *= _z; }
+void Vec3::mul(const float _x, const float _y) { this->x *= _x; this->y *= _y; }
 void Vec3::mul(const Vec3& _v, const bool _useW) { this->mul(_v.x, _v.y, _v.z, _v.w, _useW); }
-void Vec3::mul(const double _n, const bool _useW) { this->mul(_n, _n, _n, _n, _useW); }
+void Vec3::mul(const float _n, const bool _useW) { this->mul(_n, _n, _n, _n, _useW); }
 void Vec3::operator*=(const Vec3& _a) { this->mul(_a); }
-void Vec3::operator*=(const double _a) { this->mul(_a); }
+void Vec3::operator*=(const float _a) { this->mul(_a); }
 Vec3 Vec3::operator*(const Vec3& _a) const { Vec3 result = this->copy(); result.mul(_a); return result; }
-Vec3 Vec3::operator*(const double _a)   const { Vec3 result = this->copy(); result.mul(_a); return result; }
+Vec3 Vec3::operator*(const float _a)   const { Vec3 result = this->copy(); result.mul(_a); return result; }
 
 
 
-void Vec3::div(const double _x, const double _y, const double _z, const double _w, const bool _useW)
+void Vec3::div(const float _x, const float _y, const float _z, const float _w, const bool _useW)
 {
 	this->x /= _x; this->y /= _y; this->z /= _z;
 
 	if (_useW) this->w /= _w;
 }
 
-void Vec3::div(const double _x, const double _y, const double _z) { this->x /= _x; this->y /= _y; this->z /= _z; }
-void Vec3::div(const double _x, const double _y) { this->x /= _x; this->y /= _y; }
+void Vec3::div(const float _x, const float _y, const float _z) { this->x /= _x; this->y /= _y; this->z /= _z; }
+void Vec3::div(const float _x, const float _y) { this->x /= _x; this->y /= _y; }
 void Vec3::div(const Vec3& _v, const bool _useW) { this->div(_v.x, _v.y, _v.z, _v.w, _useW); }
-void Vec3::div(const double _n, const bool _useW) { this->div(_n, _n, _n, _n, _useW); }
+void Vec3::div(const float _n, const bool _useW) { this->div(_n, _n, _n, _n, _useW); }
 void Vec3::operator/=(const Vec3& _a) { this->div(_a); }
-void Vec3::operator/=(const double _a) { this->div(_a); }
+void Vec3::operator/=(const float _a) { this->div(_a); }
 Vec3 Vec3::operator/(const Vec3& _a) const { Vec3 result = this->copy(); result.div(_a); return result; }
-Vec3 Vec3::operator/(const double _a)   const { Vec3 result = this->copy(); result.div(_a); return result; }
+Vec3 Vec3::operator/(const float _a)   const { Vec3 result = this->copy(); result.div(_a); return result; }
 
 
 void Vec3::normalize() { this->div(this->getLength()); }
-void Vec3::setLength(const double _l) { this->normalize(); this->mul(_l); }
+void Vec3::setLength(const float _l) { this->normalize(); this->mul(_l); }
 
 Vec3 Vec3::copy() const { return Vec3(this->x, this->y, this->z, this->w); }
 
-double Vec3::getLength() const { return std::sqrt(std::pow(this->x, 2) + std::pow(this->y, 2) + std::pow(this->z, 2)); }
+float Vec3::getLength() const { return std::sqrt(std::pow(this->x, 2) + std::pow(this->y, 2) + std::pow(this->z, 2)); }
 
-void Vec3::constrain(const double _min, const double _max)
+void Vec3::constrain(const float _min, const float _max)
 {
 	if (this->x < _min) { this->x = _min; }
 	else if (this->x > _max) { this->x = _max; }
@@ -98,7 +98,7 @@ void Vec3::constrain(const double _min, const double _max)
 	else if (this->z > _max) { this->z = _max; }
 }
 
-void Vec3::constrainW(const double _min, const double _max)
+void Vec3::constrainW(const float _min, const float _max)
 {
 	this->constrain(_min, _max);
 
@@ -108,16 +108,16 @@ void Vec3::constrainW(const double _min, const double _max)
 
 void Vec3::intify()
 {
-	this->x = static_cast<int>(this->x);
-	this->y = static_cast<int>(this->y);
-	this->z = static_cast<int>(this->z);
+	this->x = static_cast<float>(static_cast<int>(this->x));
+	this->y = static_cast<float>(static_cast<int>(this->y));
+	this->z = static_cast<float>(static_cast<int>(this->z));
 }
 
 void Vec3::intifyW()
 {
 	this->intify();
 
-	this->w = static_cast<int>(this->w);
+	this->w = static_cast<float>(static_cast<int>(this->w));
 }
 
 Vec3 Vec3::operator*(const Mat4x4& _m) const
@@ -130,7 +130,7 @@ Vec3 Vec3::operator*(const Mat4x4& _m) const
 	);
 }
 
-void Vec3::operator*=(const Mat4x4 & _m)
+void Vec3::operator*=(const Mat4x4& _m)
 {
 	Vec3 r = this->operator*(_m);
 
@@ -138,7 +138,6 @@ void Vec3::operator*=(const Mat4x4 & _m)
 	this->y = r.y;
 	this->z = r.z;
 }
-
 
 Vec3 Vec3::intify(const Vec3& _v)
 {
@@ -156,7 +155,7 @@ Vec3 Vec3::normalize(const Vec3& _v)
 	return result;
 }
 
-double Vec3::dotProduct(const Vec3& _a, const Vec3& _b)
+float Vec3::dotProduct(const Vec3& _a, const Vec3& _b)
 {
 	return _a.x * _b.x + _a.y * _b.y + _a.z * _b.z;
 }
